backend/pyproject.toml
[project]
name = "automend-backend"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",             # async driver — used by the app at runtime
    "psycopg2-binary>=2.9.0",      # sync driver — used by alembic during migrations
    "alembic>=1.13.0",
    "pgvector>=0.3.0",

    # Redis
    "redis[hiredis]>=5.0.0",

    # Temporal
    "temporalio>=1.5.0",

    # HTTP client
    "httpx>=0.27.0",

    # Kubernetes
    "kubernetes-asyncio>=29.0.0",

    # Auth
    "pyjwt>=2.8.0",
    "passlib>=1.7.0",
    # passlib 1.7.x reads bcrypt.__about__.__version__ which was removed in
    # bcrypt 4.1+, and passlib's internal self-test trips on bcrypt 4.1's
    # stricter 72-byte enforcement. Pin to 4.0.x until passlib ships a fix or
    # we migrate off passlib. passlib is effectively unmaintained — long-term,
    # move to bcrypt-direct in `backend/app/api/routes_auth.py` and
    # `backend/scripts/bootstrap_admin.py`.
    "bcrypt>=4.0.0,<4.1.0",

    # Logging
    "structlog>=24.1.0",

    # Utilities
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "testcontainers[postgres,redis]>=4.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
# Not using strict mode — FastAPI route handlers rely on Pydantic response_model
# for type inference, which mypy --strict flags as missing annotations.
# We still catch real bugs (union-attr, arg-type, misc) at normal settings.
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
# Third-party libs with loose typing — silence Any-return noise
module = [
    "redis.*",
    "kubernetes_asyncio.*",
    "temporalio.*",
]
follow_imports = "silent"

[[tool.mypy.overrides]]
# Files that use third-party SDKs where the stubs don't match the runtime API
# (redis-py async/sync unions, kubernetes_asyncio api_client attribute, etc.)
module = [
    "app.stores.redis_store",
    "app.temporal.activities",
    "app.workers.correlation_worker",
    "app.services.classifier_server",
]
disable_error_code = ["misc", "arg-type", "attr-defined", "assignment"]
